Separation of Header files and Main files
Moved function definitions to a c file to the corresponding header file
